Allow for unconfirmed users
Make it so that we can add users who haven't signed up themselves, in
order to have a representation of users from Catarse and from our
mailing list, for example, both on our database and on Customer.io.

This will come in a further commit, but we need this change so that
when a user logs in for the first time (after getting an invitation
in the form of a sendEnrollmentEmail message), their status is
changed to that of a user of nosfiscais.